Write the bandit section of main.tex around RA-TS-F; fix int32 time-axis bug
Paper (main.tex, +~250 lines, 19 pages):
- Sec 5: add per-regime reward-distribution histograms (new fig) and a
  distributional analysis paragraph (noise-to-gap ratios, level vs shape).
- Sec 7 rewritten from placeholder: the decision problem and its five
  difficulties; policies compared (references, discounted baselines,
  TAR-UCB/LI-TAR-UCB/CG-ICLB progression); full RA-TS-F description
  (TRACK/ACQUIRE, conjugate posterior, snapshot-baseline CUSUM, level-
  invariant shape matching, coverage floor) with mechanism figure; protocol;
  one-week results (table, regret curves, internals figure, floor and
  purity-variant ablations); six-month results (table, regret/timeline
  figures, drift robustness, library-size diagnosis).
- Abstract, intro contributions, background, discussion, conclusion filled
  in; Page (1954) reference added.

Bug fix (bandit_long_experiment.py write_outputs): np.arange defaulted to
int32 on Windows, overflowing at ~97 days -- corrupted time axes in the
6-month figures and frac_optimal in summary.csv (regret/cum_reward were
unaffected). Fixed with int64; regenerated 6-month summary + figures from
checkpoints (Oracle frac_opt now 0.994, was 0.756).

Figures: reward histograms, RA-TS demo overview, week regret curves and
internals, corrected 6-month cumreward/regret and timelines copied into
tracked paper/figs/.

+++ Lightsaber_X/bandit_long_experiment.py
@@ -391,7 +391,8 @@ def write_outputs(outdir, results, cache, oracle_table, hold):
     hold_n = int(hold * fs)
     Wsec = np.asarray(cache['Wsec']); tsec = np.asarray(cache['tsec'])
     ndec = min(len(v['rewards']) for v in results.values())
-    dec_t = (np.arange(ndec) * hold_n) / fs
+    # int64: ndec*hold_n exceeds int32 beyond ~97 days (np.arange defaults to int32 on Windows)
+    dec_t = (np.arange(ndec, dtype=np.int64) * hold_n) / fs
     Wdec = np.vstack([np.interp(dec_t, tsec, Wsec[i]) for i in range(3)])     # (3, ndec)
     opt_arm = np.argmax(oracle_table @ Wdec, axis=0)                          # (ndec,)
     sev = 1.0 * Wdec[1] + 2.0 * Wdec[2]
@@ -457,7 +458,7 @@ def write_outputs(outdir, results, cache, oracle_table, hold):
     if diag_name:
         v = results[diag_name]
         fig, ax = plt.subplots(2, 1, figsize=(13, 5), sharex=True)
-        td = (np.arange(len(v['tar_J'])) * hold_n) / fs / 86400
+        td = (np.arange(len(v['tar_J']), dtype=np.int64) * hold_n) / fs / 86400
         ax[0].plot(td, v['tar_J'], 'tab:red'); ax[0].set_ylabel("library size J"); ax[0].grid(alpha=0.3)
         ax[0].set_title(f"{diag_name}: recurrent-regime library and mode")
         ax[1].plot(td, v['tar_mode'], 'tab:blue', lw=0.6)
